seperated Order into new file and fixed take order
