+++ src/CoffeeShopSimulator/ShopExampleProcedurally/Order.java
@@ -1,0 +1,10 @@
+package CoffeeShopSimulator.ShopExampleProcedurally;
+
+public class Order {
+    String guestName;
+    String order;
+    public Order(String name, String order){
+        guestName = name;
+        order = order;
+    }
+}
